+++ src/client/java/com/zurrtum/create/client/foundation/block/connected/CTSpriteShiftEntry.java
@@ -15,13 +15,13 @@ public class CTSpriteShiftEntry extends SpriteShiftEntry {
     }
 
     public float getTargetU(float localU, int index) {
-        float uOffset = (index % type.getSheetSize());
-        return getTarget().getU((getUnInterpolatedU(getOriginal(), localU) + uOffset) / ((float) type.getSheetSize()));
+        float uOffset = index % type.getSheetSize();
+        return getTarget().getU((getUnInterpolatedU(getOriginal(), localU) + uOffset) / type.getSheetSize());
     }
 
     public float getTargetV(float localV, int index) {
-        float vOffset = (index / type.getSheetSize());
-        return getTarget().getV((getUnInterpolatedV(getOriginal(), localV) + vOffset) / ((float) type.getSheetSize()));
+        float vOffset = index / type.getSheetSize();
+        return getTarget().getV((getUnInterpolatedV(getOriginal(), localV) + vOffset) / type.getSheetSize());
     }
 
 }
